+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # v1.3.0
 ## Added
 - Launch payment flow with payment method like parameter
+- Launch QR Scanner functionality
 
 # v1.2.0
 ## Added